Fix device space retrieving on reMarkable2
On the reMarkable2, the device is not /dev/mmcblk1p7, so I use the path directly in `df`. I am not able to test on the first reMarkable, but I am quite confident it will work with this version of busybox too.

I rewrote a bit the way it is retrieved too, to avoid a double SSH command, and the use of pipes in command.